fix: fix redirect to fav page
